card / diary

+++ src/main/java/com/example/myapp/context/user/Session.java
@@ -2,11 +2,10 @@ package com.example.myapp.context.user;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
-import org.springframework.stereotype.Component;
 
 @Data
 @AllArgsConstructor
@@ -15,7 +14,7 @@ import org.springframework.stereotype.Component;
 @Builder
 public class Session {
 
-  @JsonProperty("id")
-  private int id;
+    @JsonProperty("id")
+    private long id;
 
 }
